scripts/headers/requirements.txt
csnake
inflection
libclang